編集に関するテンプレートと処理を移動
